requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
Django==5.0.1
djangorestframework==3.14.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.1
idna==3.6
jieba3k==0.35.1
joblib==1.3.2
lxml==5.1.0
newspaper3k==0.2.8
nltk==3.8.1
pillow==10.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tinysegmenter==0.3
tldextract==5.1.1
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Pint==0.23
django-cors-headers==4.3.1